added middleware to handle errors

+++ controller/contactControllers.js
@@ -1,34 +1,47 @@
 const mongoose = require("mongoose");
-const { url } = require("../config/cloudinary");
+const asyncHandler = require("express-async-handler");
 const cloudinary = require("../config/cloudinary");
 const Contact = require("../models/contactModel");
+const User = require("../models/contactModel");
 
 //get all contacts
-const getContacts = async (req, res) => {
+const getContacts = asyncHandler(async (req, res) => {
   try {
-    const contacts = await Contact.find({});
+    const contacts = await Contact.find({ user_id: req.user._id });
     res.status(200).json(contacts);
   } catch (error) {
-    res.status(400).json({ msg: "request failed" });
+    res.status(400);
+    throw new Error("request failed");
   }
-};
+});
 
 //get single contact by its id
 const getContact = async (req, res) => {
   const { id } = req.params;
+
   //check if id is valid mongoose id type
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({ error: "contact not found" });
   }
+
   const contact = await Contact.findById(id);
+
   if (!contact) {
     return res.status(404).json({ error: "contact not found" });
+  }
+
+  if (!req.user) {
+    return res.status(401).json({ error: "user not found" });
+  }
+
+  if (contact.user_id !== req.user.id) {
+    return res.status(401).json({ error: "Not authorized" });
   }
   res.status(200).json(contact);
 };
 
 const createNewContact = async (req, res) => {
-  const { firstName, lastName, contactInfo, image, tag } = req.body;
+  const { firstName, lastName, contactInfo, image, category } = req.body;
   try {
     let img = { publicId: "", url: "" };
     if (image) {
@@ -43,13 +56,16 @@ const createNewContact = async (req, res) => {
         };
       }
     }
+
     const contactData = {
+      user_id: req.user._id,
       firstName,
       lastName,
       contactInfo,
       image: img,
-      tag,
+      category,
     };
+    // console.log(contactData);
     const contact = await Contact.create(contactData);
     // console.log(imgUploadRes);
     res.status(200).json(contact);
@@ -65,12 +81,21 @@ const updateContact = async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({ msg: "contact no found" });
   }
+
+  if (!req.user) {
+    return res.status(401).json({ error: "user not found" });
+  }
+
+  if (contact.user_id !== req.user.id) {
+    return res.status(401).json({ error: "Not authorized" });
+  }
+
   if (newContactImage) {
-    const destroyImageRespnse = await cloudinary.uploader.destroy(
+    const destroyImageResponse = await cloudinary.uploader.destroy(
       contact.image.publicId
     );
 
-    if (destroyImageRespnse) {
+    if (destroyImageResponse) {
       const imageUploadResponse = await cloudinary.uploader.upload(
         newContactImage,
         { upload_preset: "pally-contacts" }
@@ -86,9 +111,10 @@ const updateContact = async (req, res) => {
           image: img,
         };
 
-        const updatedContact = await Contact.findOneAndUpdate(
-          { _id: id },
-          updatedContactData
+        const updatedContact = await Contact.findByIdUpdate(
+          id,
+          updatedContactData,
+          { new: true }
         );
 
         if (!updatedContact) {
@@ -99,10 +125,9 @@ const updateContact = async (req, res) => {
     }
   } else {
     try {
-      const updatedContact = await Contact.findOneAndUpdate(
-        { _id: id },
-        contact
-      );
+      const updatedContact = await Contact.findByIdAndUpdate(id, contact, {
+        new: true,
+      });
       if (!updatedContact) {
         return res.satus(404).json({ msg: "contact not found" });
       }
@@ -119,10 +144,23 @@ const deleteContact = async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({ msg: "contact no found" });
   }
-  const contact = await Contact.findOneAndDelete({ _id: id });
+
+  const contact = await Contact.findById(id);
+
   if (!contact) {
     return res.status(404).json({ msg: "contact no found" });
   }
+
+  if (!req.user) {
+    return res.status(401).json({ error: "user not found" });
+  }
+
+  if (contact.user_id !== req.user.id) {
+    return res.status(401).json({ error: "Not authorized" });
+  }
+
+  await Contact.findOneAndDelete({ _id: id });
+
   //remove deleted contact image from cloudinary
   if (contact.image.publicId) {
     await cloudinary.uploader.destroy(contact.image.publicId);
